pyproject.toml
[project]
name = "tools_agent"
version = "0.1.0"
description = "LangGraph tools agent with MCP and a RAG tool"
authors = [
    { name = "langchain-ai" }, 
]
requires-python = ">=3.11.0,<3.13"
dependencies = [
    "langgraph>=0.6.2",
    "langchain-anthropic>=0.3.18",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "pydantic==2.11.3",
    "langchain>=0.3.27",
    "mcp>=1.23.3",
    "supabase>=2.15.1",
    "aiohttp>=3.8.0",
    "langgraph-supervisor>=0.0.11",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pyjwt>=2.8.0",
]

[tool.setuptools]
packages = ["tools_agent"]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.3.6",
    "ruff>=0.8.4",
]
